Switch to using mitol-django-scim
