+++ main/settings.py
@@ -19,6 +19,7 @@ from urllib.parse import urljoin
 
 import dj_database_url
 from django.core.exceptions import ImproperlyConfigured
+from mitol.scim.settings.scim import *  # noqa: F403
 
 from main.envs import (
     get_bool,
@@ -129,33 +130,13 @@ INSTALLED_APPS = (
     "data_fixtures",
     "vector_search",
     "ai_chat",
-    "scim",
+    "mitol.scim.apps.ScimApp",
 )
 
 if not get_bool("RUN_DATA_MIGRATIONS", default=False):
     MIGRATION_MODULES = {"data_fixtures": None}
 
-SCIM_SERVICE_PROVIDER = {
-    "SCHEME": "https",
-    # use default value,
-    # this will be overridden by value returned by BASE_LOCATION_GETTER
-    "NETLOC": "localhost",
-    "AUTHENTICATION_SCHEMES": [
-        {
-            "type": "oauth2",
-            "name": "OAuth 2",
-            "description": "Oauth 2 implemented with bearer token",
-            "specUri": "",
-            "documentationUri": "",
-        },
-    ],
-    "SERVICE_PROVIDER_CONFIG_MODEL": "scim.config.LearnSCIMServiceProviderConfig",
-    "USER_ADAPTER": "scim.adapters.LearnSCIMUser",
-    "USER_MODEL_GETTER": "scim.adapters.get_user_model_for_scim",
-    "USER_FILTER_PARSER": "scim.filters.UserFilterQuery",
-    "GET_IS_AUTHENTICATED_PREDICATE": "scim.utils.is_authenticated_predicate",
-}
-
+SCIM_SERVICE_PROVIDER["USER_ADAPTER"] = "users.adapters.LearnUserAdapter"  # noqa: F405
 
 # OAuth2TokenMiddleware must be before SCIMAuthCheckMiddleware
 # in order to insert request.user into the request.
