--- scim/config.py
@@ -1,13 +1,0 @@
-from django_scim.models import SCIMServiceProviderConfig
-
-
-class LearnSCIMServiceProviderConfig(SCIMServiceProviderConfig):
-    """Custom provider config"""
-
-    def to_dict(self):
-        result = super().to_dict()
-
-        result["bulk"]["supported"] = True
-        result["filter"]["supported"] = True
-
-        return result
