--- scim/urls.py
@@ -1,18 +1,0 @@
-"""URL configurations for profiles"""
-
-from django.urls import include, re_path
-
-from scim import views
-
-ol_scim_urls = (
-    [
-        re_path(r"^Bulk$", views.BulkView.as_view(), name="bulk"),
-        re_path(r"^Users/\.search$", views.SearchView.as_view(), name="users-search"),
-    ],
-    "ol-scim",
-)
-
-urlpatterns = [
-    re_path(r"^scim/v2/", include(ol_scim_urls)),
-    re_path(r"^scim/v2/", include("django_scim.urls", namespace="scim")),
-]
